Properly recurse over multi-case conditionals (fixes #1)
As described in #1, `if() {} else if() {} else if() {}` structures in
templates are not parsed correctly: the nodes of the first conditional
segment are recursed over, but any subsequent conditional cases in the
template are skipped (so in the JS pseudocode example above, only the
nodes within that first `if() {}` block would be visited).

The solution is to explicitly recurse over `node.els` and `node.elsif`,
so that any further conditionals or nodes nested within them will be
visited.